refactor(lb): privatize algorithms, backend, and backend_pool substrate modules
