+++ crates/lb/src/algorithms/latency_aware.rs
@@ -56,3 +56,43 @@ impl Default for LatencyAware {
         Self::new()
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use std::time::Duration;
+
+    use spooky_config::config::{Backend, HealthCheck};
+
+    use super::LatencyAware;
+    use crate::{backend::BackendState, backend_pool::BackendPool};
+
+    fn create_backend_state(address: &str, weight: u32) -> BackendState {
+        BackendState::new(&Backend {
+            id: format!("backend-{address}"),
+            address: address.to_string(),
+            weight,
+            health_check: Some(HealthCheck {
+                path: "/health".to_string(),
+                interval: 1000,
+                timeout_ms: 1000,
+                failure_threshold: 3,
+                success_threshold: 1,
+                cooldown_ms: 0,
+            }),
+        })
+    }
+
+    #[test]
+    fn latency_aware_prefers_lower_ewma() {
+        let mut pool = BackendPool::new_from_states(vec![
+            create_backend_state("10.0.0.1:1", 1),
+            create_backend_state("10.0.0.2:1", 1),
+        ]);
+
+        pool.finish_request(0, Duration::from_millis(150), Some(200));
+        pool.finish_request(1, Duration::from_millis(20), Some(200));
+
+        let mut lb = LatencyAware::new();
+        assert_eq!(lb.pick(&pool), Some(1));
+    }
+}
