+++ crates/lb/src/upstream_pool.rs
@@ -12,9 +12,9 @@ use spooky_config::runtime::{
 };
 
 use crate::{
-    backend::{BackendState, HealthTransition},
+    backend::BackendState,
     backend_pool::BackendPool,
-    health::HealthFailureReason,
+    health::{HealthFailureReason, HealthTransition},
     load_balancing::LoadBalancing,
 };
 
@@ -335,10 +335,7 @@ mod tests {
         let unhealthy = pool
             .mark_backend_failure_from_active_check(picked)
             .expect("health transition");
-        assert!(matches!(
-            unhealthy,
-            crate::backend::HealthTransition::BecameUnhealthy
-        ));
+        assert!(matches!(unhealthy, crate::HealthTransition::BecameUnhealthy));
         let after_failure = pool.membership_summary();
         assert_eq!(after_failure.total_backends, 2);
         assert_eq!(after_failure.healthy_backends, 1);
